% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities.R
\name{annotate_CTN}
\alias{annotate_CTN}
\title{Rename Consolidated CTNs Based on Dominant Cell Types}
\usage{
annotate_CTN(CTN_cluster_label, CTN_merged_celltype_prop, celltype_levs)
}
\arguments{
\item{CTN_cluster_label}{A data frame containing hierarchical clustering results for merging the highly overlapped CTNs.
It must include the following two columns: \code{CTN} (the CTN names) and
\code{cluster} (the corresponding cluster label).}

\item{CTN_merged_celltype_prop}{A data frame representing the
cell type proportion for each CTN with the following columns: \code{CTN} (the CTN names),
\code{Celltype} (cell type labels), and \code{prop} (cell type proportion of the CTN).}

\item{celltype_levs}{A character vector defining the levels or ordering of the cell types.}
}
\value{
A list contains:
\describe{
\item{annotation}{A data frame with the original CTN names (\code{CTN}),
the corresponding cluster label (\code{cluster}), and the new name for the consolidated CTN (\code{annotation}) after merging.}
\item{core_celltypes_summary}{A data frame summarizing clusters with at least 2 CTNs.
It includes the number of appearances of each core cell type in the original CTN names (\code{n_appearance}) and their relative proportion within the
consolidated CTNs (\code{prop}).}
}
}
\description{
Renaming the consolidated Cell-type Triads Niches (CTNs) using the most frequent core
cell types within the original triads. If a tie occurs, the function resolves
it by identifying the cell type with the highest overall proportion.
}
