% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkcEst.R
\name{BtwClusterSS_linear}
\alias{BtwClusterSS_linear}
\title{Computing the between-cluster sum of squares for the linear kernels}
\usage{
BtwClusterSS_linear(X, H)
}
\description{
Computes the between-cluster sum of squares for the linear kernel associated with X
}
\keyword{internal}
