% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities.R
\name{Build_cell_neighbor_maxdist}
\alias{Build_cell_neighbor_maxdist}
\title{Constructing neighborhood cell type count and proportion matrices}
\usage{
Build_cell_neighbor_maxdist(
  dat_slide,
  knn_num,
  cell_types_available,
  min.radius = 100,
  max.radius = 500,
  min.n.neighbors = 10,
  return.nn.ids = FALSE,
  prob.list = seq(0, 1, 0.25)
)
}
\arguments{
\item{dat_slide}{Input data frame for a single tissue section. Must contain columns including cell ids (\code{CellID}), spatial coordinates (\code{X}, \code{Y}), and cell type labels (\code{Celltype}).}

\item{knn_num}{The number of nearest neighbors. The default value is 20.}

\item{cell_types_available}{A vector of length \eqn{T} containing all available cell type labels.}

\item{min.radius}{Lower threshold for neighborhood distance boundary. The default value is 100.}

\item{max.radius}{Upper threshold for neighborhood distance boundary. The default value is 500.}

\item{min.n.neighbors}{Minimum number of nearest neighbors within the neighborhood distance boundary. Cells with fewer than \code{min.n.neighbors} (default: 10) neighbors will be excluded from downstream analysis.}

\item{return.nn.ids}{Whether to return cell ids of the nearest neighbors of each cell.}

\item{prob.list}{Probabilities between \eqn{[0, 1]} to produce quantiles for neighborhood distance distribution}
}
\value{
A list contains:
\describe{
\item{count.df}{An \eqn{N \times T} neighborhood cell type count matrix. Rows with all \code{NaN} values are cells with fewer than fewer than \code{min.n.neighbors} neighbors within the distance boundary.}
\item{prop.df}{An \eqn{N \times T} neighborhood cell type proportion matrix. Rows with all \code{NaN} values are cells with fewer than fewer than \code{min.n.neighbors} neighbors within the distance boundary.}
\item{dist.quantiles}{Quantiles (including the 99th percentile) of the of the distances to all nearest neighbors corresponding to the probabilities in \code{prop.list}. Can be used to examine the distribution of the distances.}
\item{neighbor.cell.ids}{A data frame containing cell ids of the nearest neighbors of each focal cell.}
}
}
\description{
Define spatial neighborhood using \code{\link[FNN:get.knn]{FNN::get.knn()}} and compute neighborhood cell counts and proportions by all available cell types.
}
\details{
The neighborhood distance boundary \code{max.dist} is initially set as the 99th percentile of all distances from focal cells to their the nearest neighbors. If it falls outside the predefined range (default: \eqn{[100, 500]}), it will be reset to the minimum (\code{min.radius}) or maximum threshold (\code{max.radius}).

For cells with fewer than \code{min.n.neighbors} neighbors within the boundary, their corresponding neighborhood cell type counts and proportions will all be set to \code{NaN} and they will be subsequently excluded from niche clustering.
}
