% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_CLQ.R
\name{compute_CLQ_observed}
\alias{compute_CLQ_observed}
\title{Computing observed colocation quotient (CLQ)}
\usage{
compute_CLQ_observed(
  cell_neighbor_counts,
  dat_in,
  min.ncell = 100,
  num_cores = 1
)
}
\arguments{
\item{cell_neighbor_counts}{\eqn{N \times T} neighborhood cell type count matrix obtained using \code{\link[=Build_cell_neighbor_maxdist]{Build_cell_neighbor_maxdist()}}.}

\item{dat_in}{Input data frame with \eqn{N} rows. Columns must contain \code{CellID}, \code{ImageID} and \code{Celltype}.}

\item{min.ncell}{Minimum cell number threshold. Tissue sections with fewer than \code{min.ncell} cells will be excluded from the analysis.}

\item{num_cores}{The number of cores to use in \code{\link[pbmcapply:pbmclapply]{pbmcapply::pbmclapply()}}, i.e. at most how many child processes will be run simultaneously. Can only be set at 1 on Windows.}
}
\value{
A data frame of \eqn{T\times T\times L} rows containing observed CLQs between focal (\code{Celltype})
and neighboring cell types (\code{Celltype_neighbor}) in each tissue section (\code{ImageID}),
where \eqn{T} is the number of available cell types and \eqn{L} is the number of tissue sections.
}
\description{
Computes the observed colocation quotient (CLQ) between available cell type pairs in each tissue section in a spatial omics dataset.
}
\details{
The colocalization of a pair of cell types \eqn{a} and \eqn{b} in tissue section \eqn{l}
can be evaluated with the observed CLQ

\deqn{CLQ_{b\rightarrow a}^{(l), obs} =
\left\{\begin{array}{ll}
0, & N_a^{(l)} \leq 5 \mbox{ or} \ N_b^{(l)} \leq 5 \\
\frac{C_{b\rightarrow a}^{(l), obs} / N_a^{(l)}}{N_b^{(l)} / (N^{(l)} - 1)}, & \text{otherwise}
\end{array}\right.}

where \eqn{C_{b\rightarrow a}^{(l), obs}} is the number of cells of cell type \eqn{b} within the neighborhood of cell type \eqn{a},
and \eqn{N_a^{(l)}}, \eqn{N_b^{(l)}}, \eqn{N^{(l)}} are the number of cells for cell type \eqn{a},
\eqn{b} and the total number of cells in tissue section \eqn{l}, respectively.
}
\references{
Bouchard, G. et al. A quantitative spatial cell-cell colocalizations framework enabling comparisons between in vitro assembloids and pathological specimens.
Nat Commun 16, 1392 (2025). \url{https://doi.org/10.1038/s41467-024-55129-6}
}
