% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_CLQ.R
\name{compute_CLQ_permuted}
\alias{compute_CLQ_permuted}
\title{Compute colocation quotient (CLQ) on permuted data}
\usage{
compute_CLQ_permuted(
  cell_neighbor_ids,
  dat_in,
  min.ncell = 100,
  num_cores = 5,
  iter.num = 500,
  seed = 42
)
}
\arguments{
\item{cell_neighbor_ids}{Cell IDs of the nearest neighbors obtained using \code{\link[=Build_cell_neighbor_maxdist]{Build_cell_neighbor_maxdist()}}.}

\item{dat_in}{An input data frame with \eqn{N} rows. Columns must contain \code{CellID}, \code{ImageID} and \code{Celltype}.}

\item{min.ncell}{An integer specifying the minimum cell count threshold. Tissue sections with fewer than \code{min.ncell} cells will be excluded from the analysis.}

\item{num_cores}{Number of cores to use in \code{\link[pbmcapply:pbmclapply]{pbmcapply::pbmclapply()}}, which determines how many permutations are
processed simultaneously. Windows environments only support \code{1} core.}

\item{iter.num}{Number of permutations (default: 500).}

\item{seed}{Random seed passed to \code{\link[base:Random]{base::set.seed()}} to ensure reproducibility when shuffling the cell type labels.}
}
\value{
A data frame of \eqn{T\times T\times L\times} \code{iter.num} rows containing the permutated CLQs between focal (\code{Celltype})
and neighboring cell types (\code{Celltype_neighbor}) in each tissue section (\code{ImageID}),
where \eqn{T} is the number of available cell types and \eqn{L} is the number of tissue sections.
}
\description{
Randomly shuffle cell type labels within each tissue section and compute the permuted CLQs between every cell type pair.

This creates a null distribution for comparison with the observed CLQ (\code{\link[=compute_CLQ_observed]{compute_CLQ_observed()}}) to compute a permutation p-value.
}
\details{
In each tissue section, the cell type labels are randomly shuffled for \code{iter.num} (default: 500) times to generate permutated CLQ values

\deqn{CLQ_{b\rightarrow a}^{(l), k} =
\left\{\begin{array}{ll}
0, & N_a^{(l)} \leq 5 \mbox{ or} \ N_b^{(l)} \leq 5 \\
\frac{C_{b\rightarrow a}^{(l), k} / N_a^{(l)}}{N_b^{(l)} / (N^{(l)} - 1)}, & \text{otherwise}
\end{array}\right.}

where \eqn{C_{b\rightarrow a}^{(l), k}} is the number of cells of cell type \eqn{b} within the neighborhood of cell type \eqn{a} in iteration \eqn{k, k=}1,2,...,\code{iter.num},
and \eqn{N_a^{(l)}}, \eqn{N_b^{(l)}}, \eqn{N^{(l)}} are the number of cells for cell type \eqn{a},
\eqn{b} and the total number of cells in tissue section \eqn{l}, respectively.
}
\references{
Bouchard, G. et al. A quantitative spatial cell-cell colocalizations framework enabling comparisons between in vitro assembloids and pathological specimens.
\emph{Nat Commun} 16, 1392 (2025). \url{https://doi.org/10.1038/s41467-024-55129-6}
}
\seealso{
\link[=Build_cell_neighbor_maxdist]{Build_cell_neighbor_maxdist},
\link[=compute_CLQ_observed]{compute_CLQ_observed},
\link[=CLQ_permutation_test]{CLQ_permutation_test}
}
