% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkcEst.R
\name{compute_H_linear}
\alias{compute_H_linear}
\title{Compute top eigenvectors (H) of the combined kernel using linear kernels}
\usage{
compute_H_linear(X_m_list, theta, centers)
}
\description{
Compute top eigenvectors (H) of the combined kernel using linear kernels
}
\keyword{internal}
