% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities.R
\name{CTN_presence_coxph}
\alias{CTN_presence_coxph}
\title{Cox regression on patient survival based on CTN presence}
\usage{
CTN_presence_coxph(
  Full_CTN_table,
  CTN_ls,
  clinical_df,
  time = "OS",
  event = "Ev.O",
  patient_id = "Patient_ID",
  confounder = NULL,
  min.n.patients = 3,
  min.n.cells = 100
)
}
\arguments{
\item{Full_CTN_table}{A data frame containing clustering labels for all cell-type triads}

\item{CTN_ls}{A character vector containing the names of the CTNs to evaluate.}

\item{clinical_df}{A data frame containing image-level clinical data. It must
include columns for patient ID (\code{patient_id}), survival time (\code{time}),
status event (\code{event}), and any specified confounders (\code{confounder}).}

\item{time}{A character string specifying the survival time column name. See \code{\link[survival:Surv]{survival::Surv()}} for details.}

\item{event}{A character string specifying the survival status indicator column name, normally 0=alive, 1=dead. See \code{\link[survival:Surv]{survival::Surv()}} for details.}

\item{patient_id}{A character string specifying the patient ID column name.}

\item{confounder}{A character string or vector specifying column names of
confounders to be adjusted for in the model. Defaults to \code{NULL}.}

\item{min.n.patients}{An integer specifying the minimum patient threshold.
Only CTNs that are both present present and absent in \eqn{\ge} \code{min.n.patients} will be evaluated. Defaults to 3.}

\item{min.n.cells}{An integer specifying the minimum cell count threshold
required to define a CTN as present on a tissue section. Defaults to 100.}
}
\value{
A data frame containing the survival analysis results for each CTN:
\describe{
\item{CTN}{The name of the CTN.}
\item{logHR}{Log harzard ratio.}
\item{HR}{Hazard ratio.}
\item{pval}{P-value}
\item{c_index}{Concordance-index (c-index)}
\item{FDR}{The Benjamini-Hochberg adjusted p-value}
}
}
\description{
Evaluates the prognostic ability of Celltype Triad Niches (CTNs) by fitting a
Cox proportional hazards regression model using patient-level CTN presence as the main predictor.
}
\details{
Patient-level presence of a CTN is defined as having \eqn{\ge} \code{min.n.cells}
cells within that niche in at least one tissue section.
}
