% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{draw_CTN_celltype}
\alias{draw_CTN_celltype}
\title{Spatial Visualization of CTNs and Cell Types}
\usage{
draw_CTN_celltype(
  Full_CTN_table,
  CTN = "Bcell_CD4T_CD8T",
  img_list,
  celltype_palette,
  CTN_palette = c(CTN = "#E69F00", Unassigned = "grey90"),
  rename_celltype = NULL,
  scales = "fixed",
  ncol = length(img_list),
  legend.ncol = NULL,
  aspect.ratio = NULL
)
}
\arguments{
\item{Full_CTN_table}{A data frame containing spatial coordinates (\code{X}, \code{Y}), image IDs (\code{ImageID}), cell type
labels (\code{Celltype}), and niche labels for all cell-type triads}

\item{CTN}{CTN label; e.g. B_CD4T_CD8T.}

\item{img_list}{A list of images to be visualized.}

\item{celltype_palette}{A named vector with colors as values and cell type labels as names. Must contain all cell types.}

\item{CTN_palette}{A named vector with colors as values and CTN labels (e.g. CTN vs. Unassigned) as names.}

\item{rename_celltype}{A function for renaming cell type labels.}

\item{scales}{\code{scales} parameter \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{ncol}{Number of columns passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{legend.ncol}{Number of columns of legend keys, passed to \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}.}

\item{aspect.ratio}{Aspect ratio of the panel, passed directly to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A list of \link[ggplot2:ggplot]{ggplot2::ggplot} objects:
\describe{
\item{p_celltype}{A spatial plot showing individual cell type distributions.}
\item{p_CTN}{A spatial plot showing the distribution of the specified CTN.}
}
}
\description{
Generates plots to visualize the spatial distribution of Celltype Triad Niches (CTNs) and individual cell types across tissue sections.

\link[ggtext:ggtext]{ggtext} package is required for the function.
}
