% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{draw_CTN_coxph_logHR}
\alias{draw_CTN_coxph_logHR}
\title{Plot Cox regression results on patient-level CTN presence}
\usage{
draw_CTN_coxph_logHR(
  coxph_res,
  CTN_merged_celltype_prop,
  top = NULL,
  alpha = 0.1,
  celltype_levs,
  celltype_palette,
  dot_fontsize = 14,
  rename_celltype = NULL,
  line_width = 0.4,
  y_text_size = NULL,
  width_ratio = c(0.5, 0.5)
)
}
\arguments{
\item{coxph_res}{A data frame representing the Cox proportional hazards regression results
on patient-level CTN presence. Output of the \code{\link[=CTN_presence_coxph]{CTN_presence_coxph()}} function}

\item{CTN_merged_celltype_prop}{A data frame representing the
cell type proportion for each CTN with the following columns: \code{CTN} (the CTN names),
\code{Celltype} (cell type labels), and \code{prop} (cell type proportion of the CTN).}

\item{top}{An integer specifying the maximum number of top-ranked CTNs ordered by \eqn{|logHR|} to display on the plot.}

\item{alpha}{A numeric value specifying the False Discovery Rate (FDR) significance threshold for filtering the CTNs.}

\item{celltype_levs}{A character vector defining the levels or ordering of the cell types.}

\item{celltype_palette}{Input of the \code{\link[=color_CTN_names]{color_CTN_names()}} function. A named vector with colors as values and cell type labels as names. Must contain all cell types.}

\item{dot_fontsize}{Input of the \code{\link[=color_CTN_names]{color_CTN_names()}} function. Size of the dots. Default is 16pt.}

\item{rename_celltype}{A function for renaming cell type labels.}

\item{line_width}{A numeric value specifying the line width for the boxes highlighting the core cell types.}

\item{y_text_size}{A numeric value specifying the font size of the y-axis labels.}

\item{width_ratio}{A numeric vector of length 2 specifying the width ratios between the log harzard ratio lollipop plot and cell type proportion bar plot.}
}
\value{
A combined \link[ggplot2:ggplot]{ggplot2::ggplot} object containing both the survival lollipop plot and the cell type composition bar plot.
}
\description{
Visualizes the prognostic value and cellular composition of Cell-type Triad Niches (CTNs).

The left panel features a lollipop plot where the x-axis represents the log hazard ratio (HR) from Cox regression models,
with lines colored by Benjamini-Hochberg adjusted P-values (red: FDR < 0.05; black: 0.05 \eqn{\le} FDR < 0.1).

The right panel features an aligned bar plot illustrating the cell type composition for each CTN,
with core cell types are highlighted with black boxes.

\link[ggtext:ggtext]{ggtext}, \link[forcats:forcats-package]{forcats}, \link[cowplot:cowplot-package]{cowplot},
and \link[patchwork:patchwork-package]{patchwork} packages are required for the function.
}
