% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{draw_CTN_dendro}
\alias{draw_CTN_dendro}
\title{Draw CTN dendrogram
Dendrogram visualization for the hierarchical clustering results on the pairwise Jaccard distance between CTNs.}
\usage{
draw_CTN_dendro(
  cl,
  h,
  k = NULL,
  CTN_annotation = NULL,
  annot_var = "annotation",
  celltype_palette = NULL,
  dot_fontsize = 12,
  CTN_group_palette = NULL,
  show_text = TRUE,
  x_expand = c(0.001, 0.001),
  y_lim = c(-2, 1),
  rename_celltype = NULL
)
}
\arguments{
\item{cl}{An object of class "hclust" which describes the tree produced by the clustering process. Output of the \code{\link[=jaccard_dist_hclust]{jaccard_dist_hclust()}} function.}

\item{h}{Numeric scalar or vector with heights where the tree should be cut. \code{\link[stats:cutree]{stats::cutree()}} for details. At least one of \code{k} or \code{h} must be specified, \code{k} overrides \code{h} if both are given.}

\item{k}{An integer scalar or vector with the desired number of groups. See \code{\link[stats:cutree]{stats::cutree()}} for details.}

\item{CTN_annotation}{(Optional) A data frame with the original CTN names (\code{CTN}), the corresponding cluster label (\code{cluster}), and the new name for the consolidated CTN (specified in \code{annot_var}) after merging.}

\item{annot_var}{Column name of the newly consolidated CTN in \code{CTN_annotation}.}

\item{celltype_palette}{Input of the \code{\link[=color_CTN_names]{color_CTN_names()}} function. A named vector with colors as values and cell type labels as names. Must contain all cell types.}

\item{dot_fontsize}{Input of the \code{\link[=color_CTN_names]{color_CTN_names()}} function. Size of the dots. Default is 16pt.}

\item{CTN_group_palette}{(Optional) A named vector with colors as values and CTN categories as names. Only used when \code{CTN_annotation} is provided.}

\item{show_text}{Whether to display CTN names alongside the dendrogram (default: \code{TRUE}).}

\item{x_expand}{A numeric vector of length two of the multiplicative range expansion factors for the x axis. See \code{\link[ggplot2:expansion]{ggplot2::expansion()}} for details.}

\item{y_lim}{A numeric vector of length two providing limits of the scale. Use NA to refer to the existing minimum or maximum.}

\item{rename_celltype}{A function for renaming cell type labels.}
}
\value{
A list with components:
\describe{
\item{p}{Dendrogram}
\item{label_df}{Label data. See \code{\link[ggdendro:dendro_data]{ggdendro::dendro_data()}} for details.}
}
}
\description{
\code{forcats}, \code{ggdendro}, \code{ggtext} packages are required for the function.
}
