% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{draw_CTN_umap}
\alias{draw_CTN_umap}
\title{Visualize CTN cell type composition similarity using UMAP}
\usage{
draw_CTN_umap(
  Full_CTN_table,
  CTN_merged_celltype_prop,
  CTN_ls,
  CTN_annotation = NULL,
  CTN_group_var = "cluster",
  CTN_group_palette = NULL,
  radius_range = c(1, 8),
  umap_n_neighbors = 10,
  umap_seed = 42,
  ...
)
}
\arguments{
\item{Full_CTN_table}{A data frame containing clustering labels for all cell-type triads}

\item{CTN_merged_celltype_prop}{A data frame representing the
cell type proportion for each CTN with the following columns: \code{CTN} (the CTN names),
\code{Celltype} (cell type labels), and \code{prop} (cell type proportion of the CTN).}

\item{CTN_ls}{A character vector containing the names of the CTNs.}

\item{CTN_annotation}{A data frame with the original CTN names (\code{CTN}), the corresponding cluster label (\code{cluster}), and the new name for the consolidated CTN (specified in \code{annot_var}) after merging.}

\item{CTN_group_var}{A character string specifying the column in \code{CTN_annotation} on CTN grouping. Will be used to color the dots.}

\item{CTN_group_palette}{A named vector with colors as values and CTN categories as names.}

\item{radius_range}{A numeric vector of length two that specifies the minimum and maximum size of the bubbles. See \code{\link[ggplot2:scale_size]{ggplot2::scale_radius()}} for details.}

\item{umap_n_neighbors}{\code{n_neighbors} parameter \code{\link[uwot:umap]{uwot::umap()}}. The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation.}

\item{umap_seed}{\code{seed} parameter \code{\link[uwot:umap]{uwot::umap()}}. Integer seed to use to initialize the random number generator state.}

\item{...}{Other parameters to pass on to \code{\link[uwot:umap]{uwot::umap()}}}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Conduct dimensionality reduction on cell-type triad niche (CTN) cell type composition using the Uniform Manifold Approximation and Projection (UMAP) method (McInnes et al., 2018).

The first two UMAP embeddings are visualized using a bubble plot, where the sizes reflect the total cell count within each CTN, and the colors denote their CTN category.

\link[ggrepel:ggrepel]{ggrepel}, \link[cowplot:cowplot-package]{cowplot}, and \link[uwot:umap]{uwot} packages are required for the function.
}
