% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{draw_htmap_jaccard_index}
\alias{draw_htmap_jaccard_index}
\title{Jaccard similarity index heatmap}
\usage{
draw_htmap_jaccard_index(jaccard_mat, cl, num_clusters)
}
\arguments{
\item{jaccard_mat}{A numeric matrix containing the Jaccard similarity indices between all pairs of CTNs. Output of the \code{\link[=jaccard_dist_hclust]{jaccard_dist_hclust()}} function.}

\item{cl}{An object of class "hclust" which describes the tree produced by the clustering process. Output of the \code{\link[=jaccard_dist_hclust]{jaccard_dist_hclust()}} function.}

\item{num_clusters}{An integer specifying the target number of clusters to slice or highlight on the heatmap.}
}
\value{
A \link[ComplexHeatmap:Heatmap-class]{ComplexHeatmap::Heatmap} object.
}
\description{
Heatmap visualization of the Jaccard similarity indices between all pairs of cell-type triad
Niches (CTNs) using an ordered heatmap based on hierarchical clustering.
}
