% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{draw_silhoutte_score}
\alias{draw_silhoutte_score}
\title{Visualize silhouette score}
\usage{
draw_silhoutte_score(hclust_silhouette, num_clusters)
}
\arguments{
\item{hclust_silhouette}{A data frame containing the average (\code{avg_si}) and median Silhouette score (\code{median_si})
under cluster number \code{K}. Output of the \code{\link[=jaccard_dist_hclust]{jaccard_dist_hclust()}} function.}

\item{num_clusters}{Optimal number of clusters.}
}
\description{
\code{cowplot} package is required for the function.
}
\details{
Evaluating hierarchical clustering performance for selecting optimal cluster number
}
