% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities.R
\name{jaccard}
\alias{jaccard}
\title{Compute Jaccard similarity index}
\usage{
jaccard(a, b)
}
\arguments{
\item{a, b}{Cell IDs of cells assigned to CTNs}
}
\value{
Jaccard similarity index
}
\description{
Compute Jaccard similarity index between two sets of niche labels.
}
