% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities.R
\name{jaccard_dist_hclust}
\alias{jaccard_dist_hclust}
\title{Perform hierarchical clustering on Jaccard distance}
\usage{
jaccard_dist_hclust(Full_CTN_table, CTN_ls, num_cores = 1)
}
\arguments{
\item{Full_CTN_table}{A data frame containing clustering labels for all cell-type triads}

\item{CTN_ls}{A character vector containing the names of the CTNs.}

\item{num_cores}{The number of cores to use in \code{\link[pbmcapply:pbmclapply]{pbmcapply::pbmclapply()}}, i.e. at most how many child processes will be run simultaneously. Can only be set at 1 on Windows.}
}
\value{
A list contains:
\describe{
\item{jaccard_mat_mgcv}{A matrix containing the Jaccard similarity indices (1-Jaccard distance) between every pair of CTNs}.
\item{cl}{An object of class "hclust" which describes the tree produced by the clustering process. See \code{\link[stats:hclust]{stats::hclust()}} for details.}
\item{hclust_silhouette}{A dataframe containing the average and median silhouette scores for evaluating cluster quality. See \code{\link[cluster:silhouette]{cluster::silhouette()}} for details.}
}
}
\description{
Computes the Jaccard distance matrix between all pairs of cell-type triad niches (CTNs) and performs hierarchical clustering analysis on the result.
}
