% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCOPE.R
\name{label_CTN}
\alias{label_CTN}
\title{Renaming CTN cluster labels}
\usage{
label_CTN(CTN_table, driver_celltypes = NULL, clusternum = 2)
}
\description{
Internal function for renaming CTN cluster labels based on the proportion of driver cell type triads
}
\keyword{internal}
