% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkcEst.R
\name{LabelToBinaryMat}
\alias{LabelToBinaryMat}
\title{Convert cluster labels to a binary cluster indicator matrix}
\usage{
LabelToBinaryMat(labels)
}
\description{
Convert cluster labels to a binary cluster indicator matrix
}
\keyword{internal}
