% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkcEst.R
\name{mkkcEst_linear}
\alias{mkkcEst_linear}
\title{Robust multiple kernel K-means clustering on a multiview data}
\usage{
mkkcEst_linear(
  X_m_list,
  centers,
  iter.max = 100,
  epsilon = 1e-04,
  theta = rep(1/length(X_m_list), length(X_m_list)),
  alpha = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{X_m_list}{A list of length \eqn{P} (default: 4) containing \eqn{P} kernel matrices.}

\item{centers}{The number of clusters \eqn{K}.}

\item{iter.max}{The maximum number of iterations allowed. The default is 100.}

\item{epsilon}{Convergence threshold. The default is \eqn{10^{-4}}.}

\item{theta}{Initial values for kernel coefficients. The default is 1/P for all views.}

\item{alpha}{Decay rate used for computing the exponential moving average of \eqn{\theta}. In interation \eqn{\scriptl}, \eqn{{\theta^{(\scriptl)}}^{EMA} = \alpha \theta^{(\scriptl)} + (1-\alpha) {\theta^{(\scriptl-1)}}^{EMA}, \alpha\in[0, 1]}. The default value is 0.5.}

\item{verbose}{Whether progress is printed during mini-batch k-means clustering. Passed on to \link[ClusterR:MiniBatchKmeans]{ClusterR::MiniBatchKmeans}. Defaults to \code{TRUE}.}
}
\value{
A list contains:
\describe{
\item{cluster}{A vector of integers (from \code{1:K}) indicating the cluster to which each point is allocated.}
\item{totss}{The total sum of squares.}
\item{withinss}{Matrix of within-cluster sum of squares by cluster, one row per view.}
\item{withinsscluster}{Vector of within-cluster sum of squares, one component per cluster.}
\item{withinssview}{Vector of within-cluster sum of squares, one component per view.}
\item{tot.withinss}{Total within-cluster sum of squares, i.e. \code{sum(withinsscluster)}.}
\item{betweenssview}{Vector of  between-cluster sum of squares, one component per view.}
\item{tot.betweenss}{The between-cluster sum of squares, i.e. \code{totss-tot.withinss}.}
\item{clustercount}{The number of clusters \code{K}.}
\item{coefficients}{The kernel coefficients}
\item{H}{The continuous clustering assignment}
\item{size}{The number of points, one component per cluster.}
\item{iter}{The number of iterations.}
\item{MbatchKm_WCSS_per_cluster}{Within-cluster sum of squares of each cluster when recovering cluster labels from the final H using minibatch K-means.}
\item{MbatchKm_centroids}{Final cluster centroids.}
}
}
\description{
An internal function adapted from MKKC::mkkcEst for performing robust multiple kernel K-means clustering on a multiview data.
}
\keyword{internal}
