% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{prepare_dendro_data}
\alias{prepare_dendro_data}
\title{Prepare input data CTN dendrogram}
\usage{
prepare_dendro_data(cl, h, k = NULL)
}
\arguments{
\item{cl}{An object of class "hclust" which describes the tree produced by the clustering process. Output of the \code{\link[=jaccard_dist_hclust]{jaccard_dist_hclust()}} function.}

\item{h}{Numeric scalar or vector with heights where the tree should be cut. \code{\link[stats:cutree]{stats::cutree()}} for details.
At least one of \code{k} or \code{h} must be specified, \code{k} overrides \code{h} if both are given.}

\item{k}{An integer scalar or vector with the desired number of groups. See \code{\link[stats:cutree]{stats::cutree()}} for details.}
}
\value{
A list with components (See \code{\link[ggdendro:dendro_data]{ggdendro::dendro_data()}} for details):
\describe{
\item{segments}{Line segment data}
\item{labels}{Label data}
}
}
\description{
Internal function for preparing input data to \code{\link[=draw_CTN_dendro]{draw_CTN_dendro()}} function

\link[ggdendro:ggdendro-package]{ggdendro} package is required for the function.
}
\references{
Atrebas. Dendrograms in R, a lightweight approach. \url{https://atrebas.github.io/post/2019-06-08-lightweight-dendrograms/}
}
\keyword{internal}
