% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCOPE.R
\name{run_SCOPE}
\alias{run_SCOPE}
\title{Multi-view cell type triad niche screening}
\usage{
run_SCOPE(
  combination_table,
  cell_neighbor_table,
  dat_in,
  clusternum = 2,
  min.prop = 0.3,
  linear = FALSE,
  mgcv_df = 15,
  iter.max = 100,
  epsilon = 1e-04,
  alpha = 0.5,
  num_cores = 1,
  rename_CTN = TRUE,
  save_results = TRUE,
  output_dir = getwd(),
  suffix = ""
)
}
\arguments{
\item{combination_table}{Data frame containing all cell type combinations for CTN screening. Each row is a combination.}

\item{cell_neighbor_table}{\eqn{N\times T} neighborhood cell type proportion table. Column names must contain all cell type labels in \code{combination_table}.}

\item{dat_in}{Input data frame with \eqn{N} rows. Columns must contain \code{CellID} and \code{Celltype}.}

\item{clusternum}{The number of clusters \eqn{K}. The default value is 2.}

\item{min.prop}{Minimum neighborhood cell type proportion threshold. Proportions below this value will all be set to 0. The default value is 0.3.}

\item{linear}{Whether to run linear version of SCOPE (i.e. not applying thin plate regression spline smoothing). Default is \code{FALSE}.}

\item{mgcv_df}{The dimension of the basis used to represent the thin plate regression spline smooth term. The default is 15. See \link[mgcv:s]{mgcv::s} for details. Disabled when \code{linear=TRUE}.}

\item{iter.max}{The maximum number of iterations allowed. The default is 100.}

\item{epsilon}{Convergence threshold. The default is \eqn{10^{-4}}.}

\item{alpha}{Decay rate \eqn{\alpha} used for computing the exponential moving average of weight vector \eqn{\theta}. In interation \eqn{l}, \eqn{{\theta^{(l)}}^{EMA} = \alpha \theta^{(l)} + (1-\alpha) {\theta^{(l-1)}}^{EMA}, \alpha\in[0, 1]}. The default value is 0.5.}

\item{num_cores}{The number of cores to use in \link[pbmcapply:pbmclapply]{pbmcapply::pbmclapply}, i.e. at most how many child processes will be run simultaneously. Can only be set at 1 on Windows.}

\item{rename_CTN}{Whether to rename niches based on the proportion of core cell types. The cluster with the highest proportion will be named as \code{CTN} when \code{clusternum = 2} and \code{CTN1} otherwise. The cluster with the lowest core cell type proportion will be renamed as \code{Unassigned}. Other clusters will be named as \code{CTN2, ...} accordingly when \code{clusternum > 2}.}

\item{save_results}{Whether to save the niche screening results to the output directory. If \code{TRUE} (default) then the results for each CTN will be saved into a separate file. If \code{FALSE} then will all CTN screening results will be combined into a single data frame and returned as function output.}

\item{output_dir}{Output directory. Will automatically create a new directory when the target directory doesn't exist.}

\item{suffix}{File name suffix.}
}
\value{
\describe{
\item{res}{CTN screening results. Only returns when \code{save_results} is \code{FALSE}.}
}
}
\description{
Cell type triad niche (CTN) detection using SCOPE.
}
