% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCOPE.R
\name{run_SCOPE_subcluster}
\alias{run_SCOPE_subcluster}
\title{Subclustering cells labeled as "CTN" in initial analysis}
\usage{
run_SCOPE_subcluster(
  Full_CTN_table,
  cell_neighbor_table,
  CTN_annotation,
  annot_var = "annotation",
  subclusternum = 2,
  output_dir = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{Full_CTN_table}{A data frame containing spatial coordinates (\code{X}, \code{Y}), image IDs (\code{ImageID}), cell type
labels (\code{Celltype}), and niche labels for all cell-type triads}

\item{cell_neighbor_table}{Same cell_neighbor_table used as the input for the initial clustering}

\item{CTN_annotation}{Data frame with CTN annotations. Must contain columns \code{CTN} (initial CTN names) and \code{annot_var} (consolidated CTN names)}

\item{annot_var}{A character string specicying the column name representing the consolidated CTN names in \code{CTN_annotation}}

\item{subclusternum}{Number of subclusters}

\item{output_dir}{Output directory used in \code{\link[=run_SCOPE]{run_SCOPE()}}.}

\item{verbose}{Whether progress is printed during mini-batch k-means clustering. Passed on to \link[ClusterR:MiniBatchKmeans]{ClusterR::MiniBatchKmeans}. Defaults to \code{TRUE}.}
}
\value{
A list with the following attributes:
\describe{
\item{Full_CTN_table_reclustered}{A data frame containing the subcluster labels with the same structure as \code{Full_CTN_table}}
\item{WCSS}{Within-cluster sum of squares}
}
}
\description{
Subclustering cells labeled as "CTN" in initial analysis
}
