% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkcEst.R
\name{StandardizeData}
\alias{StandardizeData}
\title{Interal function for standardizing each data matrix X_m}
\usage{
StandardizeData(X, center = TRUE, scale = TRUE)
}
\description{
Interal function for standardizing each data matrix X_m
}
\keyword{internal}
