% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkcEst.R
\name{WithinClusterSS_linear}
\alias{WithinClusterSS_linear}
\title{Computing the within-cluster sum of squares for the linear kernels}
\usage{
WithinClusterSS_linear(X, H)
}
\description{
Computes the within-cluster sum of squares for the linear kernel associated with X
}
\keyword{internal}
