% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOPE_Utilities_figures.R
\name{draw_CTN_celltype_prop}
\alias{draw_CTN_celltype_prop}
\title{CTN cell type proportion dotplot}
\usage{
draw_CTN_celltype_prop(
  CTN_merged_celltype_prop,
  CTN_dend,
  facet_var = "clust",
  celltype_palette,
  radius_range = c(0.5, 8),
  rename_celltype = waiver(),
  x_expand = c(0.05, 0.05)
)
}
\arguments{
\item{CTN_merged_celltype_prop}{A data frame representing the
cell type proportion for each CTN with the following columns: \code{CTN} (the CTN names),
\code{Celltype} (cell type labels), and \code{prop} (cell type proportion of the CTN).}

\item{CTN_dend}{Output of the \code{\link[=draw_CTN_dendro]{draw_CTN_dendro()}} function}

\item{facet_var}{Faceting variables, usually CTN category. Default to \code{"clust"} from \code{CTN_dend$label_df}}

\item{celltype_palette}{A named vector with colors as values and cell type labels as names. Must contain all cell types.}

\item{radius_range}{A numeric vector of length two that specifies the minimum and maximum size of the bubbles. See \code{\link[ggplot2:scale_size]{ggplot2::scale_radius()}} for details.}

\item{rename_celltype}{A function for renaming cell type labels.}

\item{x_expand}{A numeric vector of length two of the multiplicative range expansion factors for the x axis. See \code{\link[ggplot2:expansion]{ggplot2::expansion()}} for details.}
}
\description{
\code{cowplot} and \code{ggtext} packages are required for the function.
}
\details{
Dotplot visualization of cell type abundance within each cell-type triad niche (CTN).

Dotplot visualization of the cell type abundance of cell-type triad niches (CTN). Bubbles are colored by cell type labels, and their sizes represent relative cell type proportions with each CTN.
}
